some refactoring and removal of test data arrays from components
