+++ apps/garage-admin/src/types/adminTypes.ts
@@ -1,0 +1,17 @@
+import { User, Thing } from '@my-garage/common';
+
+export type BookingData = {
+  user: User;
+  thing: Thing;
+  date: Date;
+};
+
+export type BookingsElements = {
+  [current: string]: JSX.Element;
+  previous: JSX.Element;
+};
+
+export type TabList = {
+  key: string;
+  tab: string;
+};
