+++ packages/eslint-config-garage/frontend.js
@@ -4,5 +4,7 @@ module.exports = {
     'react/jsx-one-expression-per-line': 0,
     'react/react-in-jsx-scope': 0,
     '@typescript-eslint/indent': 'off',
+    'react/jsx-props-no-spreading': 'off',
+    'react/jsx-wrap-multilines': 'off',
   },
 };
